requirements.txt
flask
langchain
langchain-community
sentence-transformers
pypdf
chromadb
huggingface_hub
